PacketTunnel: extract types into separate files and re-arrange files
