+++ ios/PacketTunnel/MullvadEndpoint+WgEndpoint.swift
@@ -1,0 +1,22 @@
+//
+//  MullvadEndpoint+WgEndpoint.swift
+//  PacketTunnel
+//
+//  Created by pronebird on 15/07/2022.
+//  Copyright © 2022 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+import WireGuardKit
+
+extension MullvadEndpoint {
+    var ipv4RelayEndpoint: Endpoint {
+        return Endpoint(host: .ipv4(ipv4Relay.ip), port: .init(integerLiteral: ipv4Relay.port))
+    }
+
+    var ipv6RelayEndpoint: Endpoint? {
+        guard let ipv6Relay = ipv6Relay else { return nil }
+
+        return Endpoint(host: .ipv6(ipv6Relay.ip), port: .init(integerLiteral: ipv6Relay.port))
+    }
+}
