+++ ios/PacketTunnel/PacketTunnelProvider.swift
@@ -415,109 +415,3 @@ class PacketTunnelProvider: NEPacketTunnelProvider, TunnelMonitorDelegate {
         )
     }
 }
-
-struct PacketTunnelConfiguration {
-    var tunnelSettings: TunnelSettingsV2
-    var selectorResult: RelaySelectorResult
-}
-
-extension PacketTunnelConfiguration {
-    var wgTunnelConfig: TunnelConfiguration {
-        let mullvadEndpoint = selectorResult.endpoint
-        var peers = [mullvadEndpoint.ipv4RelayEndpoint]
-        if let ipv6RelayEndpoint = mullvadEndpoint.ipv6RelayEndpoint {
-            peers.append(ipv6RelayEndpoint)
-        }
-
-        let peerConfigs = peers.compactMap { (endpoint) -> PeerConfiguration in
-            let pubKey = PublicKey(rawValue: selectorResult.endpoint.publicKey)!
-            var peerConfig = PeerConfiguration(publicKey: pubKey)
-            peerConfig.endpoint = endpoint
-            peerConfig.allowedIPs = [
-                IPAddressRange(from: "0.0.0.0/0")!,
-                IPAddressRange(from: "::/0")!
-            ]
-            return peerConfig
-        }
-
-        var interfaceConfig = InterfaceConfiguration(
-            privateKey: tunnelSettings.device.wgKeyData.privateKey
-        )
-        interfaceConfig.listenPort = 0
-        interfaceConfig.dns = dnsServers.map { DNSServer(address: $0) }
-        interfaceConfig.addresses = [
-            tunnelSettings.device.ipv4Address,
-            tunnelSettings.device.ipv6Address
-        ]
-
-        return TunnelConfiguration(name: nil, interface: interfaceConfig, peers: peerConfigs)
-    }
-
-    var dnsServers: [IPAddress] {
-        let mullvadEndpoint = selectorResult.endpoint
-        let dnsSettings = tunnelSettings.dnsSettings
-
-        if dnsSettings.effectiveEnableCustomDNS {
-            let dnsServers = dnsSettings.customDNSDomains
-                .prefix(DNSSettings.maxAllowedCustomDNSDomains)
-            return Array(dnsServers)
-        } else {
-            if let serverAddress = dnsSettings.blockingOptions.serverAddress {
-                return [serverAddress]
-            } else {
-                return [mullvadEndpoint.ipv4Gateway, mullvadEndpoint.ipv6Gateway]
-            }
-        }
-    }
-}
-
-extension WireGuardLogLevel {
-    var loggerLevel: Logger.Level {
-        switch self {
-        case .verbose:
-            return .debug
-        case .error:
-            return .error
-        }
-    }
-}
-
-extension WireGuardAdapterError: LocalizedError {
-    public var errorDescription: String? {
-        switch self {
-        case .cannotLocateTunnelFileDescriptor:
-            return "Failure to locate tunnel file descriptor."
-
-        case .invalidState:
-            return "Failure to perform an operation in such state."
-
-        case .dnsResolution(let resolutionErrors):
-            let detailedErrorDescription = resolutionErrors
-                .enumerated()
-                .map { index, dnsResolutionError in
-                    return "\(index): \(dnsResolutionError.address) \(dnsResolutionError.errorDescription ?? "???")"
-                }
-                .joined(separator: "\n")
-
-            return "Failure to resolve endpoints:\n\(detailedErrorDescription)"
-
-        case .setNetworkSettings:
-            return "Failure to set network settings."
-
-        case .startWireGuardBackend(let code):
-            return "Failure to start WireGuard backend (error code: \(code))."
-        }
-    }
-}
-
-extension MullvadEndpoint {
-    var ipv4RelayEndpoint: Endpoint {
-        return Endpoint(host: .ipv4(ipv4Relay.ip), port: .init(integerLiteral: ipv4Relay.port))
-    }
-
-    var ipv6RelayEndpoint: Endpoint? {
-        guard let ipv6Relay = ipv6Relay else { return nil }
-
-        return Endpoint(host: .ipv6(ipv6Relay.ip), port: .init(integerLiteral: ipv6Relay.port))
-    }
-}
