+++ ios/PacketTunnel/WireGuardLogLevel+Logging.swift
@@ -1,0 +1,22 @@
+//
+//  WireGuardLogLevel+Logging.swift
+//  PacketTunnel
+//
+//  Created by pronebird on 15/07/2022.
+//  Copyright © 2022 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+import WireGuardKit
+import Logging
+
+extension WireGuardLogLevel {
+    var loggerLevel: Logger.Level {
+        switch self {
+        case .verbose:
+            return .debug
+        case .error:
+            return .error
+        }
+    }
+}
